0.3.0 - Removing unnecessary functionality in Cyclops Engine.
Removed exception wrapping.  This functionality was added at the tail-end of a previous project and I can't justify it for general use.  If some type of semi-global exception handling is needed, just wrap the engine from the outside.

Removing sentinel code.  The sentinel was used to maintain proper order for CyclopsRoutine status reports that could be queried from an arbitrary position within the update loop.  It wasn't needed.  Just query from the outside.