0.5.0
Added optional pooling functionality to CyclopsRoutine to reduce GC pressure.  It's there when you need it, easy to use, and will stay out of your way.  All built-in CyclopsRoutines use it.  In the case of multiple CyclopsEngine instances running on different threads, everything should remain thread-safe as well.  This isn't fully tested, but all existing test cases pass.  TODO: Although not required for most scenarios, providing adjustable maximum capacities for pooling and the ability to clear the pools will likely be a good thing for someone somewhere, so it's on the list.
Added several tweening helper structs to simplify the creation of tweening oriented CyclopsRoutines.
Added a material tweening smoke test that uses a Standard Material because Standard Materials are compatible (for testing purposes) with other pipelines.
Added a smoke test to sanity check pooling in the sense that no exceptions are thrown when a stressful version of typical activity (that would trigger pooling) is encountered across multiple frames.
Changed all built-in CyclopsRoutines.  They all use pooling.
Changed tweening oriented CyclopsRoutines.  They all use the new helper structs.
Removed CyclopsIterate.  It was unlikely to ever be used.
Removed various built-in CyclopsRoutine overloads including anything using a propertyName string instead of a propertyId integer.