Removed default tags. Improved timers.
Default tags have a few quality of life benefits in testing and debugging.  That said, they're really not required and remain mostly unused.  Tags can be added as needed for testing.  Debugging will likely be updated to provide a view of both tags and class names.
Timers were simplified and are now correctly removed when tag based removals are requested.  This was past behavior that didn't make it into the new framework until just now.
TODO: Timers should be affected by pausing as well.  They were when they weren't special cased as they are now.